refactor: renamed engine_connector file to simulation_connector
