+++ src/rai_sim/rai_sim/o3de/o3de_connector.py
@@ -24,7 +24,7 @@ from geometry_msgs.msg import Pose
 from rai.communication.ros2.connectors import ROS2ARIConnector, ROS2ARIMessage
 from tf2_geometry_msgs import do_transform_pose
 
-from rai_sim.engine_connector import (
+from rai_sim.simulation_connector import (
     Entity,
     PoseModel,
     SceneState,
